fix: Auth cache not clearing after auth request. (#51)
The auth cache is not being cleaned correctly after a request is complete. Either for success or failure. This gives stale auth responses when the server has been running long enough to need to get a refresh token, but there is no token available since it is using the old cached request.

fixes #50